agents/pyproject.toml
[project]
name = "gnucash-agents"
version = "0.1.0"
description = "AI Agents service for GnuCash with Strands framework and MCP integration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.3",
    "fastapi>=0.128.1",
    "mcp>=1.26.0",
    "openai>=1.0.0",
    "pydantic-settings>=2.12.0",
    "strands-agents>=1.24.0",
    "strands-agents-tools>=0.2.19",
]
